backend/requirements.txt
flask
flask-cors
textblob
openai
python-dotenv
